Moves Radio and Input sizes to external files to omit the build warnings
