+++ src/components/Form/Input/Sizes.ts
@@ -1,0 +1,3 @@
+import { Size } from './Size';
+
+export type Sizes = keyof typeof Size;
