+++ src/components/Form/Radio/Size.ts
@@ -1,5 +1,3 @@
-export type Sizes = keyof typeof Size;
-
 export enum Size {
   sm = 'input-sm',
   lg = 'input-lg',
